Add method security: Owner handling
Enable @GlobalMethodSecurity and set up a custom ExpressionEvaluator
that can handles domain object ownership.

The OwnerPermissionEvaluator uses svenson to detect a "ownerId" property
and expects it to be a foreign key to app_user.id.

null ownerIds can be either allowed or throw an error.

The storeFoo() mutation has now a special access control annotation

    @PreAuthorize(
"(hasRole('ROLE_USER') and hasPermission(#foo, 'dummy')) or hasRole('ROLE_ADMIN')"
    )
    public boolean storeFoo(Foo foo)

So we always allow access if the user has ROLE_ADMIN and if it has
ROLE_USER and fulfills the hasPermission() which routes the given
Foo object to our expression evaluator

The current database is set up so that there are two users
"user-a"/"user-b" with the same passwords.

"user-a" owns 'Foo Aaa' and 'user-b' owns 'Foo B'.

So if you use the badly secured Foo Admin interface as one of those
users you should be able to edit and succesfully save one of them,
but not the other.