Solve GraphQL auth conundrum
So we have an internal GraphQL service point which we always POST to and
as such it falls under Spring securities' CSRF protection as well as
general authentication rules, which is what we want.

On the other hand, we want to be able to auto-complete queries in the
IDE and maybe execute some queries while we write code.

This requires skipping CSRF protection though as the IntelliJ GraphQL
plugin doesn't know about CSRF.

We could disable CSRF protection optionally on the normal end point, but
it seems better to have two different end points:

One which is always CSRF-protected and which is used by the application
and one that is optionally present to accept non-CSRF requests from the
IDE.

This is what this commit does.

It also changes our graphql.config.json config file to point to the new
/graphql-dev URL.